inforcetest/urls.py
"""
URL configuration for inforcetest project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/5.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView, TokenVerifyView

from auth.views import RegisterView
from restaurant.views import Menulist,RestaurantListAPI
from users.views import VoteListView,ResultListView
urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/token/verify/', TokenVerifyView.as_view(), name='token_verify'),
    path('menus/',Menulist.as_view(),name='all'),
    path('menus/<int:pk>/',Menulist.as_view(),name='all'),
    path('vote/',VoteListView.as_view(),name='vote'),
    path('vote/<int:pk>/',VoteListView.as_view(),name='vote'),
    path('restaurant/',RestaurantListAPI.as_view(),name='restaurant'),
    path('restaurant/<int:pk>/',RestaurantListAPI.as_view(),name='restaurant'),
    path('result/',ResultListView.as_view(),name='result'),
    path('register/',RegisterView.as_view(),name='register'),

]
