requirements.txt
Django~=5.2.7
djangorestframework~=3.16.1
pip~=23.2.1
pytest~=8.4.2
asgiref~=3.10.0
sqlparse~=0.5.3
psycopg2-binary~=2.9.11
Pygments~=2.19.2
pluggy~=1.6.0
iniconfig~=2.3.0
packaging~=25.0
PyJWT~=2.10.1
djangorestframework-simplejwt>=5.3.1